Make changes to support continuous integration This includes linting enforcement during CI builds.
One linting violation currently in the code is the "TribeList_id" field in the TribeList schema. MongoDB automatically includes a "_id" field in all collections, which follows a different convention than our linting style guide.  However, these _id fields do not need to be explicitly specified in the Mongoose schema (and probably shouldn't be, to avoid possibly naming-synchronization errors) in order to be used.  So we can address the linting issue by simply removing the field from the Mongoose schema.

In addition, renamed TribesListSchema to TribeListSchema to conform to the Mongoose convention of using singular nouns for schema names.